Extract reusable SoundCloudCrate component from SoundCloudLibrary
Pull the crate-detail view (track table + analyze-on-play FIFO queue +
in-crate search/sort + embedded Widget player) out of SoundCloudLibrary
into a new reusable SoundCloudCrate component. SoundCloudLibrary keeps the
crate grid and now delegates the detail view to SoundCloudCrate.

No behavior change — this is the foundation for the unified library
(Phase 1b), where the combined grid will open SoundCloud crates through
this same component. SoundCloudCrate takes crate/token/backend/
onRefreshToken/onBack props and is self-contained (its own scFetch +
analysis queue), so it can be mounted from anywhere.
